\name{advRlab3-package}
\alias{advRlab3-package}
\alias{advRlab3}
\docType{package}
\title{
\packageTitle{advRlab3}
}
\description{
\packageDescription{advRlab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{advRlab3}
\packageIndices{advRlab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{advRlab3}

Maintainer: \packageMaintainer{advRlab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
